Add auto-complete
Add function auto_complete_longest_common_prefix(). This functions
overwrites default TAB behaviour. If option --auto-complete is used
the variable auto_complete_func is set to a function pointer. This
function will be called on keypress TAB. The function calculates the
maximum common prefix and updates the input field to that value. Any
leading white space charcters will also be removed.

This function is fully aware of the case-insensitive option as it only
uses the fstrncmp() function provided by dmenu.